Frontend and Backend
Backend: mapped a new update method which updates votes on a certain post, also database has a post table which contains all posts.
Frontend: Added a new post component which is integrated inside popular component and will also be integrated in other components that's why it's in shared folder. made a post service in which i have created 2 methods for getting posts and updating post votes.